+++ zc/include/ZC_rw.h
@@ -37,16 +37,16 @@ void ZC_symTransform_8bytes(unsigned char data[8]);
 void ZC_symTransform_2bytes(unsigned char data[2]);
 void ZC_symTransform_4bytes(unsigned char data[4]);
 
-int ZC_checkFileSize(char *srcFilePath);
-unsigned char *ZC_readByteData(char *srcFilePath, int *byteLength);
-double *ZC_readDoubleData_systemEndian(char *srcFilePath, int *nbEle);
-float *ZC_readFloatData_systemEndian(char *srcFilePath, int *nbEle);
-double *ZC_readDoubleData(char *srcFilePath, int *nbEle);
-float *ZC_readFloatData(char *srcFilePath, int *nbEle);
-void ZC_writeByteData(unsigned char *bytes, int byteLength, char *tgtFilePath);
-void ZC_writeDoubleData(double *data, int nbEle, char *tgtFilePath);
-void ZC_writeFloatData(float *data, int nbEle, char *tgtFilePath);
-void ZC_writeData(void *data, int dataType, int nbEle, char *tgtFilePath);
+size_t ZC_checkFileSize(char *srcFilePath);
+unsigned char *ZC_readByteData(char *srcFilePath, size_t *byteLength);
+double *ZC_readDoubleData_systemEndian(char *srcFilePath, size_t *nbEle);
+float *ZC_readFloatData_systemEndian(char *srcFilePath, size_t *nbEle);
+double *ZC_readDoubleData(char *srcFilePath, size_t *nbEle);
+float *ZC_readFloatData(char *srcFilePath, size_t *nbEle);
+void ZC_writeByteData(unsigned char *bytes, size_t byteLength, char *tgtFilePath);
+void ZC_writeDoubleData(double *data, size_t nbEle, char *tgtFilePath);
+void ZC_writeFloatData(float *data, size_t nbEle, char *tgtFilePath);
+void ZC_writeData(void *data, int dataType, size_t nbEle, char *tgtFilePath);
 
 int ZC_writeStrings(int string_size, char **string, char *tgtFilePath);
 
@@ -54,18 +54,18 @@ StringLine* createStringLineHeader();
 StringLine* createOneStringLine(char* lineContent);
 StringLine* appendOneLine(StringLine* tail, char* str);
 StringLine* ZC_readLines(char* filePath, int *lineCount);
-int ZC_writeLines(StringLine* lineHeader, char *tgtFilePath);
+size_t ZC_writeLines(StringLine* lineHeader, char *tgtFilePath);
 void ZC_replaceLines(StringLine* originalLines, char* matchKeyword, char* replaceString);
-int ZC_insertLines(char* keyAnnotationLine, StringLine* globalLineHeader, StringLine* toAddLineHeader);
+size_t ZC_insertLines(char* keyAnnotationLine, StringLine* globalLineHeader, StringLine* toAddLineHeader);
 void ZC_appendLines(StringLine* globalLineHeader, StringLine* toAddLineHeader);
 
 void ZC_freeLines(StringLine* header);
-void ZC_freeCharArrayLines(char** lines, int lineNum);
+void ZC_freeCharArrayLines(char** lines, size_t lineNum);
 
 char* rmFileExtension(char* fullFileName);
 
-void ZC_writeFloatData_inBytes(float *data, int nbEle, char* tgtFilePath);
-void ZC_writeDoubleData_inBytes(double *data, int nbEle, char* tgtFilePath);
+void ZC_writeFloatData_inBytes(float *data, size_t nbEle, char* tgtFilePath);
+void ZC_writeDoubleData_inBytes(double *data, size_t nbEle, char* tgtFilePath);
 
 #ifdef __cplusplus
 }
